Move paycheck-related elements to the 'paycheck' element group folder
