+++ angular/components/elements/element-group-paycheck/paycheck-bill-fields/paycheck-bill-fields.component.scss
@@ -1,0 +1,10 @@
+// paycheck-bill-fields.component.scss
+@import "../../../../styles/global-variables";
+@import "../../../../styles/global-mixins";
+@import "../../../../styles/global-functions";
+@import "../../../../styles/global-placeholders";
+
+paycheck-bill-fields {
+    @include flex(space-around, center, row);
+    width: 100%;
+}
