+++ angular/components/elements/element-group-paycheck/paycheck-bill-fields/paycheck-bill-fields.component.ts
@@ -1,0 +1,19 @@
+// paycheck-bill-fields.component.ts 
+import {Component, Input} from "@angular/core";
+import {Bill} from "../../../../models/bill/Bill";
+import {PaycheckFieldType} from "../paycheck/models/PaycheckFieldType";
+import {OneTimeBill, RecurringBill} from "../../../../models/paycheck/types";
+
+@Component({
+    selector: 'paycheck-bill-fields',
+    templateUrl: './paycheck-bill-fields.component.html',
+    styleUrls: ['./paycheck-bill-fields.component.css']
+})
+export class PaycheckBillFieldsComponent {
+    @Input() protected bill: OneTimeBill | RecurringBill;
+    constructor() {
+        
+    }
+
+    protected readonly PaycheckFieldType = PaycheckFieldType;
+}
