+++ angular/components/elements/element-group-paycheck/paycheck-bill/paycheck-bill.component.scss
@@ -1,0 +1,23 @@
+// paycheck-bill.component.scss
+@import "../../../../styles/global-variables";
+@import "../../../../styles/global-mixins";
+@import "../../../../styles/global-functions";
+@import "../../../../styles/global-placeholders";
+
+paycheck-bill {
+    @include flex(flex-start, center, row);
+    margin: $three-quarter-percent 0;
+    width: 95%;
+    background-color: $faded-neutral-black;
+    transition: background-color 0.11s ease-in-out;
+    padding: $half-percent;
+    @extend %soft-border-radius;
+    .bill-name {
+        text-decoration: underline;
+        border-right: gray $thin-border solid;
+        width: $twenty-percent;
+    }
+    &:hover {
+        background-color: $mint;
+    }
+}
