+++ angular/components/elements/element-group-paycheck/paycheck-income-fields/paycheck-income-fields.component.scss
@@ -1,0 +1,10 @@
+// paycheck-income-fields.component.scss
+@import "../../../../styles/global-variables";
+@import "../../../../styles/global-mixins";
+@import "../../../../styles/global-functions";
+@import "../../../../styles/global-placeholders";
+
+paycheck-income-fields {
+    @include flex(space-around, center, row);
+    width: 100%;
+}
