+++ angular/components/elements/element-group-paycheck/paycheck-income-list/paycheck-income-list.component.scss
@@ -1,0 +1,10 @@
+// paycheck-income-list.component.scss
+@import "../../../../styles/global-variables";
+@import "../../../../styles/global-mixins";
+@import "../../../../styles/global-functions";
+@import "../../../../styles/global-placeholders";
+
+paycheck-income-list {
+    width: 100%;
+    @include flex(center, center, column);
+}
