+++ angular/components/elements/element-group-paycheck/paycheck-pay-period/paycheck-pay-period.component.scss
@@ -1,0 +1,16 @@
+// paycheck-pay-period.component.scss
+@import "../../../../styles/global-variables";
+@import "../../../../styles/global-mixins";
+@import "../../../../styles/global-functions";
+@import "../../../../styles/global-placeholders";
+
+paycheck-pay-period {
+    @include flex(center, center, row);
+    width: 100%;
+    .paycheck-title-text-div {
+        margin-right: $half-percent;
+    }
+    .paycheck-pay-period {
+        text-decoration: underline;
+    }
+}
