--- angular/components/elements/paycheck-bill-field/paycheck-bill-field.component.scss
@@ -1,15 +1,0 @@
-// paycheck-bill-field.component.scss
-@import "../../../styles/global-variables";
-@import "../../../styles/global-mixins";
-@import "../../../styles/global-functions";
-@import "../../../styles/global-placeholders";
-
-paycheck-bill-field {
-    @include flex(flex-start, center, column);
-    margin: 0 $half-percent;
-    &.bill-text {
-        subtitle {
-            color: red;
-        }
-    }
-}
