--- angular/components/elements/paycheck-bill-list/paycheck-bill-list.component.scss
@@ -1,10 +1,0 @@
-// paycheck-bill-list.component.scss
-@import "../../../styles/global-variables";
-@import "../../../styles/global-mixins";
-@import "../../../styles/global-functions";
-@import "../../../styles/global-placeholders";
-
-paycheck-bill-list {
-    width: 100%;
-    @include flex(center, center, column);
-}
