--- angular/components/elements/paycheck-total/paycheck-total.component.scss
@@ -1,33 +1,0 @@
-// paycheck-total.component.scss
-@import "../../../styles/global-variables";
-@import "../../../styles/global-mixins";
-@import "../../../styles/global-functions";
-@import "../../../styles/global-placeholders";
-
-paycheck-total {
-    @include flex(center, center, column);
-    width: 100%;
-    margin: $half-percent 0;
-    subtitle {
-        color: green;
-    }
-    &.income {
-        subtitle {
-            color: green;
-        }
-    }
-    &.bill {
-        subtitle {
-            color: red;
-        }
-    }
-    &.leftover {
-        background-color: $mint;
-        @extend %soft-border-radius;
-        padding: $one-percent;
-        width: auto;
-        subtitle, ss-title {
-            color: $navy-blue;
-        }
-    }
-}
